added quotes page and moved menu-categories into sub-pages
