+++ src/sub-pages/Quotes.js
@@ -1,0 +1,14 @@
+const styles = {
+    center: {
+        textAlign: 'center'
+    }
+}
+function Quotes(){
+    return (
+        <>
+        <h1 style={styles.center}>Quote of the Day</h1>
+        </>
+    )
+}
+
+export default Quotes
